ci: use tox from pip, not OS
It was intended to use python 3.9 in this test configuration.
Using tox as packaged by the OS does not accomplish that, as that tox
will instead use the system's python (3.8) which propagates through to
the testenv.

Drop the apt-get install of tox and go with pip all the way.

Should fix the failure observed at
https://github.com/release-engineering/django-kobo-exporter/pull/34.